Revision - Major/minor - step 6

+++ wjs/plugins/wjs_submission/step6/views.py
@@ -1,18 +1,58 @@
+import logging
+
 from core import models as core_models
-from core.models import File
 from django.db.models import QuerySet
-from django.http import HttpRequest
 from django.urls import reverse_lazy
 from django.views.generic import DeleteView, DetailView, FormView, UpdateView
-from django_q.tasks import async_task
-from events import logic as event_logic
 from submission.models import Article
 from utils.setting_handler import get_setting
 
-from .. import settings as submission_settings
-from ..management.commands.send_feedback import Command as FakeYakunin
 from ..mixins import AuthorFilteringView, HtmxMixin, StepCheckView
-from .forms import SubmissionStep6Form, UploadArticleForm
+from ..models import RevisionStorage
+from ..workflow import get_feedback_ws_url, is_revision, is_revision_confirm, is_revision_full, is_revision_metadata
+from .forms import RevisionStep6Form, RevisionUploadArticleForm, SubmissionStep6Form, UploadArticleForm
+
+logger = logging.getLogger(__name__)
+
+
+def get_files(article: Article) -> dict:
+    """
+    Return a dictionary of files querysets.
+
+    This is intended to be used to update the context data of step6 and step8 views
+    and those of Upload/Delete files views.
+    """
+    files_by_type = {}
+    if is_revision(article):
+        data = RevisionStorage.objects.get(article=article).data
+
+        if file_id := data["manuscript_files"]:
+            files_by_type["manuscript_files"] = core_models.File.objects.filter(id=file_id)
+        elif file_id := data["source_files"]:
+            # Show source file if manuscript conversion failed
+            files_by_type["manuscript_files"] = core_models.File.objects.filter(id=file_id)
+        else:
+            files_by_type["manuscript_files"] = core_models.File.objects.none()
+
+        # TODO specs#2330: review data-figure vs supplementary vs administrative files relation
+        files_by_type["supplementary_files"] = core_models.File.objects.filter(
+            id__in=data["supplementary_files"],
+        )
+        files_by_type["data_figure_files"] = core_models.File.objects.filter(
+            id__in=data["data_figure_files"],
+        )
+        files_by_type["administrative_files"] = core_models.File.objects.filter(
+            id__in=data["administrative_files"],
+        )
+
+    else:
+        files_by_type["manuscript_files"] = (
+            article.manuscript_files if article.manuscript_files.exists() else article.source_files
+        )
+        files_by_type["data_figure_files"] = article.data_figure_files
+        files_by_type["administrative_files"] = article.submission_data.administrative_files
+
+    return files_by_type
 
 
 class SubmissionStep6View(AuthorFilteringView, StepCheckView, UpdateView):
@@ -20,6 +60,22 @@ class SubmissionStep6View(AuthorFilteringView, StepCheckView, UpdateView):
     step = 6
     form_class = SubmissionStep6Form
     template_name = "wjs_submission/step6/article_form.html"
+
+    def get_form_class(self):
+        """
+        Return the form class to use based on whether this is a revision.
+
+        :return: Form class to use.
+        :rtype: django.forms.Form
+        """
+        if is_revision_confirm(self.object):
+            raise ValueError("Revisions with flow-type confirm-previous-version should never touch the files!")
+        if is_revision_metadata(self.object):
+            raise ValueError("Revisions with flow-type metadata-change should never touch the files!")
+        if is_revision_full(self.object):
+            return RevisionStep6Form
+
+        return SubmissionStep6Form
 
     def get_success_url(self):
         """
@@ -74,6 +130,10 @@ class SubmissionStep6View(AuthorFilteringView, StepCheckView, UpdateView):
         """
         context = super().get_context_data(**kwargs)
         context["feedback_ws_url"] = get_feedback_ws_url(self.request, self.object.pk, self.request.user.pk)
+
+        # Include files (manuscript_files, data_figure_files, etc.)
+        context.update(get_files(article=self.object))
+
         return context
 
 
@@ -90,26 +150,26 @@ class TableRenderingContext:
         :param kwargs: Additional keyword arguments passed to the method.
         :return: The modified context dictionary with additional article files and related attributes.
         :rtype: dict
-        :raises Article.DoesNotExist: If the article with the provided `article_id` does not exist.
         """
         context = super().get_context_data(**kwargs)
         context["article"] = self._article
-        if self.kwargs["file_type"] == "manuscript":
+
+        file_type = self.kwargs.get("file_type")
+        context["button_name"] = f"trigger_{file_type}"
+        context["file_type"] = file_type
+
+        files_by_type = get_files(self._article)
+        if file_type == "manuscript":
+            context["files_list"] = files_by_type["manuscript_files"]
             context["show_conversion"] = True
-            context["files_list"] = (
-                context["article"].manuscript_files
-                if context["article"].manuscript_files.exists()
-                else context["article"].source_files
-            )
             context["failed_conversion_log"] = core_models.File.objects.filter(
-                article_id=context["article"].pk, label="Failed conversion log ConvertManuscriptToPdf"
+                article_id=self._article.pk, label="Failed conversion log ConvertManuscriptToPdf"
             ).first()
-        elif self.kwargs["file_type"] == "data":
-            context["files_list"] = context["article"].data_figure_files
-        elif self.kwargs["file_type"] == "administrative":
-            context["files_list"] = context["article"].submission_data.administrative_files
-        context["button_name"] = f"trigger_{self.kwargs['file_type']}"
-        context["file_type"] = self.kwargs["file_type"]
+        elif file_type == "data":
+            context["files_list"] = files_by_type["data_figure_files"]
+        elif file_type == "administrative":
+            context["files_list"] = files_by_type["administrative_files"]
+
         return context
 
 
@@ -121,7 +181,7 @@ class RenderSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext
 
 
 class DeleteSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext, DeleteView):
-    model = File
+    model = core_models.File
     pk_url_kwarg = "file_id"
     template_name = "wjs_submission/step6/includes/files_table.html"
 
@@ -140,17 +200,72 @@ class DeleteSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext
         """
         Delete the selected file.
 
-        If file to delete is a source file, manuscript files are cleared.
-        Same if the file is a manuscript file, source files are cleared.
+        - If file to delete is a source file, manuscript files are cleared.
+        - If file to delete is a manuscript file, source files are cleared.
+        - The file object itself is deleted.
+
+        For revisions:
+        - Only delete file objects that are NOT in the article's existing file slots.
+        - Files from the original article are preserved; only their references in RevisionStorage are removed.
+        - Newly uploaded files (not in original article) are deleted completely.
         """
-        if self.object in self._article.source_files.all():
-            for f in self._article.manuscript_files.all():
-                f.delete()
-        if self.object in self._article.manuscript_files.all():
-            for f in self._article.source_files.all():
-                f.delete()
-        self.object.delete()
-        self._article.refresh_from_db()
+        if not is_revision(self._article):
+            if self.object in self._article.source_files.all():
+                for f in self._article.manuscript_files.all():
+                    f.delete()
+            if self.object in self._article.manuscript_files.all():
+                for f in self._article.source_files.all():
+                    f.delete()
+            self.object.delete()
+            self._article.refresh_from_db()
+        else:
+            revision_storage = RevisionStorage.objects.get(article=self._article)
+            file_type = self.kwargs["file_type"]
+
+            if file_type == "manuscript":
+                # I'ts possible that the author deletes the "manuscript", before the conversion is fihished.
+                # Ensure that the object id is either the manuscript or source slot,
+                # then delete all existing Files refrenced by revision storage manuscript or source slots.
+                #
+                # below, "filter(bool...)" is needed because "empty" slots hold the "Null" value,
+                # which makes for a valid item in a list, but not for a valid id in a query
+                # (the "set" is not strictly necessary :)
+                fileids_to_delete = set(
+                    filter(bool, [revision_storage.data["manuscript_files"], revision_storage.data["source_files"]])
+                )
+                if self.object.id in fileids_to_delete:
+                    revision_storage.data["manuscript_files"] = None
+                    revision_storage.data["source_files"] = None
+                    for f in core_models.File.objects.filter(id__in=fileids_to_delete):
+                        f.delete()
+                else:
+                    logger.error(
+                        f"Unexpected file to delete {self.object.id} not manuscript nor source"
+                        f" for article {self._article.id}",
+                    )
+
+            elif file_type == "data":
+                revision_storage.data["data_figure_files"].remove(self.object.id)
+                if self.object.id not in set(self._article.data_figure_files.values_list("id", flat=True)):
+                    self.object.delete()
+
+            elif file_type == "supplementary_files":
+                # TODO specs#2330: ⚠ supplementary files are not core.File, but core.SupplementaryFiles!
+                revision_storage.data["supplementary_files"].remove(self.object.id)
+                if self.object.id not in set(self._article.supplementary_files.values_list("id", flat=True)):
+                    self.object.delete()
+
+            elif file_type == "administrative":
+                revision_storage.data["administrative_files"].remove(self.object.id)
+                if self.object.id not in set(
+                    self._article.submission_data.administrative_files.values_list("id", flat=True)
+                ):
+                    self.object.delete()
+
+            else:
+                logger.error(f"""Trying to delete unexpected file type "{file_type}" for article {self._article.id}""")
+
+            revision_storage.save()
 
     def form_valid(self, form):
         """
@@ -167,19 +282,20 @@ class DeleteSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext
 
 class UploadSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext, FormView):
     """
-    A view to allow an author to upload files during the submission of a revision.
+    A view to allow an author to upload files during the submission or a revision.
 
     Uploaded files can be manuscript, data-figure files and cover letter file.
 
     This view is intended to be used from inside a small modal.
     """
 
+    # HELP: IIC, we are mimicing a DetailView, but I don't see the gain (and this confuses me...)
     model = Article
     pk_url_kwarg = "article_id"
-    form_class = UploadArticleForm
     render_table = False
     """
-    Flag to signal whether ot render the upload form or the full files table
+    Flag to signal whether ot render the upload form or the full files table.
+    See also step4.views.ModalRenderingMixin
     """
 
     def setup(self, request, *args, **kwargs):
@@ -226,6 +342,18 @@ class UploadSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext
             ).process_value()
         return initial
 
+    def get_form_class(self):
+        """
+        Return the form class to use based on whether this is a revision.
+
+        :return: Form class to use.
+        :rtype: django.forms.Form
+        """
+        if is_revision(self.object):
+            return RevisionUploadArticleForm
+
+        return UploadArticleForm
+
     def get_form_kwargs(self):
         """
         Retrieve additional keyword arguments for initializing a form.
@@ -244,46 +372,20 @@ class UploadSubmissionFile(HtmxMixin, AuthorFilteringView, TableRenderingContext
         kwargs["file_type"] = self.file_type
         kwargs["instance"] = self.object
         kwargs["user"] = self.request.user
+        kwargs["request"] = self.request
         return kwargs
 
     def form_valid(self, form):
         """If the form is valid, save the file and return a response."""
         form.save()
-        new_file = form.new_file
-        if new_file and self.file_type == "manuscript":
-            article_id = self.object.pk
-            user_id = self.request.user.pk
-            feedback_ws_name = get_feedback_ws_name(article_id, user_id)
-            feedback_ws_url = get_feedback_ws_url(self.request, article_id, user_id)
-            event_logic.Events.raise_event(
-                event_logic.Events.ON_ARTICLE_FILE_UPLOAD,
-                request=self.request,
-                file_id=new_file,
-                original_filename=new_file.original_filename,
-                file_type="manuscript:async",
-                article=self.object,
-                feedback_ws_url=feedback_ws_url,
-                feedback_ws_name=feedback_ws_name,
-            )
-            if submission_settings.SIMULATE_YAKUNIN:
-                async_task(simulate_yakunin_call, feedback_ws_name, task_name="simulate-feedback")
         self.render_table = True
         response = self.render_to_response(self.get_context_data(form=form))
         response.headers["HX-Retarget"] = f"#files_table_{self.file_type}"
         response.headers["HX-Trigger-After-Swap"] = "closeModal"
         return response
 
-
-def get_feedback_ws_name(workflow_pk: int, user_pk: int) -> str:
-    """Compute a paper/user/situation unique name for the feedback channel."""
-    return f"submission-{workflow_pk}-{user_pk}"
-
-
-def get_feedback_ws_url(request: HttpRequest, workflow_pk: int, user_pk: int) -> str:
-    """Compute the full URL of the websocket feedback consumer."""
-    feedback_ws_name = get_feedback_ws_name(workflow_pk, user_pk)
-    return f"{'wss' if request.is_secure() else 'ws'}://{request.get_host()}/ws/feedback/{feedback_ws_name}/"
-
-
-def simulate_yakunin_call(ws_name):
-    FakeYakunin().handle(ws_name=ws_name)
+    def form_invalid(self, form):
+        """Save form and redirect via HTMX."""
+        response = super().form_invalid(form)
+        response["HX-Retarget"] = "#htmxModalContent"
+        return response
